Construct peername properly, also for IPv6
Constructing the peername naively as `ip:port` breaks IPv6 support on
newer versions of NET-SNMP (at least 5.9, maybe 5.8). For IPv6
addresses, port number is interpreted as part of the IP address. IPv6
addresses should be enclosed in brackets to separate them from port
number specifications.

This used to work, but the functionality was disabled by
4372e67c4ce3987abf3dc231ea8fde7932ad8a00

This fixes Uninett/pynetsnmp#2